requirements.txt
acres==0.5.0
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.19.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
attrs==25.4.0
bcrypt==3.2.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
databases==0.9.0
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
etelemetry==0.3.1
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
filelock==3.20.0
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
looseversion==1.3.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.3.5
oauthlib==3.3.1
openai==2.14.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathlib==1.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
propcache==0.4.1
prov==2.1.1
psycopg2-binary==2.9.11
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.6
PyMuPDFb==1.23.7
pyparsing==3.2.5
pypdfium2==5.7.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.5.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
scipy==1.16.3
sendgrid==6.12.5
sentry-sdk==2.47.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.50.0
stripe==14.1.0
tqdm==4.67.1
traits==7.0.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
